requirements.txt
black==22.3.0
eth-ape>=0.6.7
pysha3==1.0.2